updating the expression view and badge counter
