+++ Shared/UI/Views/SymbolImageView.swift
@@ -23,6 +23,17 @@ class SymbolImageView: UIImageView {
         super.init(coder: coder)
     }
     
+    func setPoint(size: CGFloat) {
+        let pointCofig = UIImage.SymbolConfiguration(pointSize: size)
+        
+        if let current = self.preferredSymbolConfiguration {
+            current.applying(pointCofig)
+            self.preferredSymbolConfiguration = current
+        } else {
+            self.preferredSymbolConfiguration = pointCofig
+        }
+    }
+    
     func set(symbol: ImageSymbol, configuration: UIImage.SymbolConfiguration? = nil) {
 
         self.symbol = symbol
